+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - ⚙️ Button UI component now exports an object instead of the element directly.
 - ⚙️ Checkbox UI component now exports an object instead of the element directly.
 - ⚙️ FieldsContainer UI component now exports an object instead of the element directly.
+- ⚙️ ItemCheckbox UI component now exports an object instead of the a fragment.
 - ⚙️ Unified checkbox naming (sometimes it was "checkbox" and others it was "checkBox").
 - ⚙️ Replaced syntax error throw when datasheet set.data is not an array with a type error throw.
 - ⚙️ Improved fieldset component's default legend text.